+++ apps/fluent-tester/src/FluentTester/TestComponents/NativeDatePicker/index.ts
@@ -1,0 +1,2 @@
+export * from './NativeDatePickerTest';
+export * from './consts';