+++ packages/components/Checkbox/src/checkmark/checkmark.macos.ts
@@ -1,0 +1,2 @@
+import checkmarkSvg from '../assets/checkmark_mac.svg';
+export default checkmarkSvg;