+++ packages/components/ContextualMenu/src/SubmenuItem.settings.ts
@@ -7,7 +7,7 @@ export const settings: IComposeSettings<SubmenuItemType> = [
       backgroundColor: 'menuBackground',
       color: 'menuItemText',
       borderColor: 'transparent',
-      borderWidth: 1
+      borderWidth: 1,
     },
     root: {
       accessible: true,
@@ -15,25 +15,34 @@ export const settings: IComposeSettings<SubmenuItemType> = [
       focusable: true,
       style: {
         display: 'flex',
-        alignItems: 'flex-start',
         flexDirection: 'row',
         alignSelf: 'flex-start',
-        width: '100%'
-      }
+        width: '100%',
+        justifyContent: 'space-between',
+      },
     },
     content: {},
-    icon: {},
-    stack: {
+    icon: {style: { marginEnd: 5 }},
+    leftstack: {
       style: {
         display: 'flex',
-        paddingStart: 16,
-        paddingEnd: 16,
+        paddingStart: 5,
         alignItems: 'center',
         flexDirection: 'row',
         alignSelf: 'flex-start',
         minHeight: 32,
-        minWidth: 80,
-        justifyContent: 'center'
+        justifyContent: 'flex-start',
+      }
+    },
+    rightstack: {
+      style: {
+        display: 'flex',
+        paddingEnd: 5,
+        alignItems: 'center',
+        flexDirection: 'row',
+        minHeight: 32,
+        width: 12,
+        justifyContent: 'flex-end',
       }
     },
     _precedence: ['focused', 'hovered', 'pressed', 'disabled'],
@@ -42,13 +51,13 @@ export const settings: IComposeSettings<SubmenuItemType> = [
         tokens: {
           backgroundColor: 'menuBackground',
           color: 'disabledText',
-        }
+        },
       },
       pressed: {
         tokens: {
           backgroundColor: 'menuItemBackgroundHovered',
           color: 'menuItemTextHovered',
-        }
+        },
       },
       focused: {
         tokens: {
@@ -58,21 +67,21 @@ export const settings: IComposeSettings<SubmenuItemType> = [
         _overrides: {
           disabled: {
             tokens: {
-              borderColor: 'focusBorder'
-            }
+              borderColor: 'focusBorder',
+            },
           },
           hovered: {
             _overrides: {
               disabled: {
                 tokens: {
-                  borderColor: 'transparent'
-                }
-              }
-            }
-          }
-        }
-      }
-    }
+                  borderColor: 'transparent',
+                },
+              },
+            },
+          },
+        },
+      },
+    },
   },
-  submenuItemName
+  submenuItemName,
 ];