+++ packages/theming/apple-theme/src/appleColors.macos.ts
@@ -492,5 +492,13 @@ export function fallbackApplePalette(): ThemeColorDefinition {
     buttonDisabledSecondaryContent: fluentUIApple.brandForegroundDisabled,
     buttonHoveredSecondaryContent: fluentUIApple.neutralInverted,
     buttonPressedSecondaryContent: fluentUIApple.neutralInverted,
+
+    checkboxBackground: fluentUIApple.communicationBlue,
+    checkboxBackgroundDisabled: fluentUIApple.brandBackgroundDisabled,
+    checkboxBorderColor: fluentUIApple.gray600,
+    checkmarkColor: fluentUIApple.neutralInverted,
+
+    personaActivityRing: fluentUIApple.neutralInverted,
+    personaActivityGlow: fluentUIApple.red10,
   };
 }