+++ packages/components/PersonaCoin/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/persona-coin
 
-This log was last generated on Tue, 15 Jun 2021 00:55:36 GMT and should not be manually modified.
+This log was last generated on Fri, 25 Jun 2021 06:21:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.4
+
+Fri, 25 Jun 2021 06:21:11 GMT
+
+### Patches
+
+- Add activity ring to PersonaCoin (tamasane@gmail.com)
+
+## 0.8.3
+
+Wed, 23 Jun 2021 18:53:56 GMT
+
+### Patches
+
+- Change persona coin to pull colors from tokens (ruaraki@microsoft.com)
+
+## 0.8.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to components, run (ruaraki@microsoft.com)
 
 ## 0.8.0