+++ packages/experimental/Button/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/experimental-button
 
-This log was last generated on Tue, 15 Jun 2021 00:55:36 GMT and should not be manually modified.
+This log was last generated on Fri, 25 Jun 2021 19:39:16 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.0
+
+Fri, 25 Jun 2021 19:39:16 GMT
+
+### Minor changes
+
+- Add chevron to submenu (lehon@microsoft.com)
+
+## 0.6.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to experimental, run (ruaraki@microsoft.com)
 
 ## 0.6.0