+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Tue, 15 Jun 2021 00:55:36 GMT and should not be manually modified.
+This log was last generated on Thu, 24 Jun 2021 21:55:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.4
+
+Thu, 24 Jun 2021 21:55:11 GMT
+
+### Patches
+
+- revert change that breaks icon on macOS (sanajmi@microsoft.com)
+
+## 0.7.2
+
+Mon, 21 Jun 2021 11:19:28 GMT
+
+### Patches
+
+- Update checkbox settings/theme for android (tamasane@gmail.com)
+
+## 0.7.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to components, run (ruaraki@microsoft.com)
 
 ## 0.7.0