+++ packages/theming/android-theme/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/android-theme
 
-This log was last generated on Tue, 15 Jun 2021 00:55:36 GMT and should not be manually modified.
+This log was last generated on Fri, 25 Jun 2021 06:21:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.0
+
+Fri, 25 Jun 2021 06:21:11 GMT
+
+### Minor changes
+
+- Improved test page on Android. (tamasane@gmail.com)
+
+## 0.3.2
+
+Mon, 21 Jun 2021 11:19:28 GMT
+
+### Patches
+
+- Update checkbox settings/theme for android (tamasane@gmail.com)
+
+## 0.3.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to libraries core and theming, run (ruaraki@microsoft.com)
 
 ## 0.3.0