+++ packages/theming/theme-tokens/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/theme-tokens
 
-This log was last generated on Tue, 15 Jun 2021 00:55:36 GMT and should not be manually modified.
+This log was last generated on Wed, 23 Jun 2021 18:53:56 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.2
+
+Wed, 23 Jun 2021 18:53:56 GMT
+
+### Patches
+
+- Fix tokens (ruaraki@microsoft.com)
+
+## 0.2.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to libraries core and theming, run (ruaraki@microsoft.com)
 
 ## 0.2.0